rental invoice refactor, number generator, migrations, added test profile

+++ src/main/java/com/example/nesta/payment/api/VerifyPayload.java
@@ -6,6 +6,6 @@ public record VerifyPayload(
         String sessionId,
         int amount,
         String currency,
-        int orderId,
+        long orderId,
         String sign
 ) {}
